requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=./requirements/requirements.txt requirements/development.in
#
alembic==1.4.1            # via flask-migrate
appdirs==1.4.3            # via black
attrs==19.3.0             # via black, pytest
black==19.10b0            # via -r requirements/development.in
click==7.0                # via black, flask
entrypoints==0.3          # via flake8
flake8==3.7.9             # via -r requirements/development.in
flask-migrate==2.5.2      # via -r requirements/base.in
flask-sqlalchemy==2.4.1   # via -r requirements/base.in, flask-migrate
flask==1.1.1              # via -r requirements/base.in, flask-migrate, flask-sqlalchemy
importlib-metadata==1.5.0  # via pluggy, pytest
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via flask
mako==1.1.2               # via alembic
markupsafe==1.1.1         # via jinja2, mako
mccabe==0.6.1             # via flake8
more-itertools==8.2.0     # via pytest
packaging==20.1           # via pytest
pathspec==0.7.0           # via black
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.4    # via -r requirements/development.in
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pyparsing==2.4.6          # via packaging
pytest==5.3.5             # via -r requirements/development.in
python-dateutil==2.8.1    # via alembic
python-dotenv==0.12.0     # via -r requirements/base.in
python-editor==1.0.4      # via alembic
regex==2020.2.20          # via black
six==1.14.0               # via packaging, python-dateutil
sqlalchemy==1.3.13        # via alembic, flask-sqlalchemy
toml==0.10.0              # via black
typed-ast==1.4.1          # via black
wcwidth==0.1.8            # via pytest
werkzeug==1.0.0           # via flask
zipp==3.0.0               # via importlib-metadata
